fs_mark: Add more sync methods for better tunability
Add a couple of more sync methods for mor refined testing. Specifically,
the following actions were taken:

1. -S7 - Here, we fsync the file after write_file() is completed but
	 only close the files once all writes are complete.

2. -S8 - Here, we only write the file in the main loop. Then, we loop
	 over open files and fysnc then and finally, we loop again and
	 close them.

Mainly, the new options focus on when the fsync and closing of the files
takes place. These were added keeping preallocation benchamrking in mind
as when the files are sync'd and closed actually affects how
preallocations behave.

Signed-off-by: Ojaswin Mujoo <ojaswin@linux.ibm.com>